requirements.txt
streamlit>=1.40.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
faiss-cpu>=1.9.0
pypdf>=5.0.1
numpy>=1.26.0
datasets>=3.0.0
ragas>=0.2.0
